simplifies running tests, docker-compose to provision 5 different database containers, fixes #98
