+++ test/docker-compose.yaml
@@ -1,0 +1,104 @@
+version: "3.4"
+services:
+  postgres:
+    image: "postgres"
+    ports:
+      - "5432:5432"
+    environment:
+      - POSTGRES_PASSWORD=supersecret
+      - POSTGRES_DB=migrator
+    volumes:
+      - ./create-test-tenants.sql:/docker-entrypoint-initdb.d/create-test-tenants.sql
+  pgadmin:
+    image: "thajeztah/pgadmin4"
+    depends_on:
+      - postgres
+    ports:
+      - "5050:5050"
+    links:
+      - postgres
+  mysql:
+    image: "mysql"
+    ports:
+      - "3306:3306"
+    environment:
+      - MYSQL_ROOT_PASSWORD=supersecret
+    volumes:
+      - ./create-test-tenants.sql:/docker-entrypoint-initdb.d/create-test-tenants.sql
+  mariadb:
+    image: "mariadb"
+    ports:
+      - "13306:3306"
+    environment:
+      - MYSQL_ROOT_PASSWORD=supersecret
+    volumes:
+      - ./create-test-tenants.sql:/docker-entrypoint-initdb.d/create-test-tenants.sql
+  percona:
+    image: "percona"
+    ports:
+      - "23306:3306"
+    environment:
+      - MYSQL_ROOT_PASSWORD=supersecret
+    volumes:
+      - ./create-test-tenants.sql:/docker-entrypoint-initdb.d/create-test-tenants.sql
+  phpmyadmin:
+    image: "phpmyadmin/phpmyadmin"
+    depends_on:
+      - mysql
+      - mariadb
+      - percona
+    ports:
+      - "8888:80"
+    environment:
+      - PMA_HOSTS=mysql,mariadb,percona
+      - PMA_USER=root
+      - PMA_PASSWORD=supersecret
+    links:
+      - mysql
+      - mariadb
+      - percona
+  mssql:
+    image: mcr.microsoft.com/mssql/server:2017-latest
+    ports:
+      - "1433:1433"
+    environment: 
+      - SA_PASSWORD=Super5ecret
+      - ACCEPT_EULA=Y
+    volumes:
+      - ./create-test-tenants-mssql.sql:/docker-entrypoint-initdb.d/create-test-tenants-mssql.sql
+    command:
+      - /bin/bash
+      - -c 
+      - |
+        /opt/mssql/bin/sqlservr &
+        PID=$$!
+        is_up=-1
+        while [ $$is_up -ne 0 ] && [ $$is_up -ne 16 ] ; do 
+          /opt/mssql-tools/bin/sqlcmd -l 30 -S localhost -h-1 -V1 -U sa -P $$SA_PASSWORD -Q "CREATE DATABASE migrator"
+          is_up=$$?
+          sleep 5
+        done
+        if [ $$is_up -eq 0 ]; then
+          for script in /docker-entrypoint-initdb.d/*.sql
+            do /opt/mssql-tools/bin/sqlcmd -U sa -P $$SA_PASSWORD -d migrator -l 30 -e -i $$script
+          done
+        fi
+        wait $$PID
+  migrator:
+    image: "lukasz/migrator"
+    depends_on:
+      - mysql
+      - mariadb
+      - percona
+      - postgres
+    ports:
+      - "8181:8080"
+    environment:
+      - MIGRATOR_YAML=/data/migrator-docker.yaml
+    volumes:
+      - .:/data
+    links:
+      - mysql
+      - mariadb
+      - percona
+      - postgres
