--- test/docker/scripts/destroy-container.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-function destroy_container() {
-  name=$1
-  docker stop "migrator-$name"
-  docker rm "migrator-$name"
-}
