+++ test/migrator-percona.yaml
@@ -1,6 +1,6 @@
 baseLocation: test/migrations
 driver: mysql
-dataSource: "root:supersecret@tcp(127.0.0.1:3306)/migrator?parseTime=true&timeout=1s"
+dataSource: "root:supersecret@tcp(127.0.0.1:23306)/migrator?parseTime=true&timeout=1s"
 singleMigrations:
   - ref
   - config
